backend/requirements.txt
backcall==0.1.0
bleach==3.0.2
certifi==2018.11.29
cycler==0.10.0
Cython==0.29.2
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.5
django-cors-headers==2.2.0
djangorestframework==3.8.2
entrypoints==0.3
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.3.0
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pipenv==2018.10.13
prometheus-client==0.5.0
prompt-toolkit==2.0.7
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.9
pyzmq==17.1.2
qtconsole==4.4.3
Send2Trash==1.5.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
virtualenv==16.0.0
virtualenv-clone==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
gunicorn==19.6.0